[refactor] 사용하지 않는 라이브러리 삭제 및 패키지 구조 수정
