+++ server/src/main/java/com/moovda_project/moovda/movie/entity/genre/Genre.java
@@ -1,10 +1,10 @@
 package com.moovda_project.moovda.movie.entity.genre;
 
-import lombok.AccessLevel;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.data.domain.PageRequest;
+
 
 import javax.persistence.*;
