+++ server/src/main/java/com/moovda_project/moovda/movie/entity/staff/MovieStaff.java
@@ -1,14 +1,12 @@
 package com.moovda_project.moovda.movie.entity.staff;
 
 import com.moovda_project.moovda.movie.entity.Movie;
-import lombok.AccessLevel;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @NoArgsConstructor()
