+++ server/src/main/java/com/moovda_project/moovda/movie/entity/watch/ToWatch.java
@@ -1,7 +1,8 @@
-package com.moovda_project.moovda.movie.entity;
+package com.moovda_project.moovda.movie.entity.watch;
 
 import com.moovda_project.moovda.audit.Auditable;
 import com.moovda_project.moovda.member.entity.Member;
+import com.moovda_project.moovda.movie.entity.Movie;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,14 +13,14 @@ import javax.persistence.*;
 @Entity
 @Getter
 @Setter
-@NoArgsConstructor()
-@Table(name = "watched")
-public class Watched extends Auditable {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "to_watch")
+public class ToWatch extends Auditable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "watched_id")
-    private Long watchedId;
+    @Column(name = "to_watch_id")
+    private Long toWatchId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "movie_id")
@@ -31,8 +32,8 @@ public class Watched extends Auditable {
 
     public void setMember(Member member) {
         this.member = member;
-        if(!this.member.getWatchedList().contains(this)) {
-            this.member.addWatchedList(this);
+        if(!this.member.getToWatchList().contains(this)) {
+            this.member.addToWatchList(this);
         }
     }
 }
