+++ server/src/main/java/com/moovda_project/moovda/movie/repository/staff/StaffRepository.java
@@ -1,5 +1,6 @@
-package com.moovda_project.moovda.movie.entity.staff;
+package com.moovda_project.moovda.movie.repository.staff;
 
+import com.moovda_project.moovda.movie.entity.staff.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StaffRepository extends JpaRepository<Staff,Long> {
