+++ server/src/main/java/com/moovda_project/moovda/movie/repository/watch/ToWatchRepository.java
@@ -1,6 +1,6 @@
-package com.moovda_project.moovda.movie.repository;
+package com.moovda_project.moovda.movie.repository.watch;
 
-import com.moovda_project.moovda.movie.entity.ToWatch;
+import com.moovda_project.moovda.movie.entity.watch.ToWatch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ToWatchRepository extends JpaRepository<ToWatch, Long> {
