+++ server/src/main/java/com/moovda_project/moovda/movie/repository/watch/WatchedRepository.java
@@ -1,6 +1,6 @@
-package com.moovda_project.moovda.movie.repository;
+package com.moovda_project.moovda.movie.repository.watch;
 
-import com.moovda_project.moovda.movie.entity.Watched;
+import com.moovda_project.moovda.movie.entity.watch.Watched;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WatchedRepository extends JpaRepository<Watched,Long> {
